commit 20230808_07-54
видалив консольлоги  та створив винес функцію створення розмітки картки  в файл markup.js
